unit test workflow

+++ config/keep_documentation.py
@@ -2,9 +2,10 @@ import os
 import shutil
 
 working_directory = r"./"
-retain = ["config", ".github", "src"]
+retain = ["config", ".github", ".git", "src"]
 os.chdir(working_directory)
 
 for item in next(os.walk(working_directory))[1]:
-    shutil.rmtree(item)
+    if item not in retain:
+        shutil.rmtree(item)
